requirements.txt
openai>=1.50
pydantic>=2.8
python-dotenv>=1.0
numpy>=1.26
pandas>=2.2
matplotlib>=3.9
seaborn>=0.13
scikit-learn>=1.4
tqdm>=4.66
tenacity>=8.4
diskcache>=5.6
jupyter>=1.0
ipykernel>=6.29
torch>=2.4
transformers>=4.44
datasets>=2.20
accelerate>=0.33
pyarrow>=16
